+++ ProgettoIDM/src/main/java/com/progetto/fr_vagoni/FRPasseggeri.java
@@ -13,7 +13,7 @@ import com.progetto.vagoni.PostoASedere;
 @DiscriminatorValue("PASSEGGERI_FR")
 public class FRPasseggeri extends Passeggeri {
 
-	public FRPasseggeri(Integer id, String marca, double peso, double lunghezza, Colore colore,
+	public FRPasseggeri(Long id, String marca, double peso, double lunghezza, Colore colore,
 			Integer numeroMassimoPasseggeri, Integer numeroPasseggeri) {
 		super(id, marca, peso, lunghezza, colore, numeroMassimoPasseggeri, numeroPasseggeri);
 	}
