+++ ProgettoIDM/src/main/java/com/progetto/tn_vagoni/TNLocomotiva.java
@@ -10,7 +10,7 @@ import com.progetto.vagoni.Locomotiva;
 @DiscriminatorValue("LOCOMOTIVA_TN")
 public class TNLocomotiva extends Locomotiva {
 
-	public TNLocomotiva(Integer id, String marca, double peso, double lunghezza, Colore colore,
+	public TNLocomotiva(Long id, String marca, double peso, double lunghezza, Colore colore,
 			Integer potenzaMotore, Integer carburanteMassimo, Integer carburanteAttuale) {
 		super(id, marca, peso, lunghezza, colore, potenzaMotore, carburanteMassimo, carburanteAttuale);
 	}
